calendarItemDetails: cap fetchedBy
This is a hacky way to show a friendly name for Magister, the only service that
provides sH with calendarItems. Magister's  friendly name is just it's normal
name capitalized.

This hack is needed since the client can only
get the friendly name of a service by calling a method, which is unhandy to
every time the details are opened.
We should have a way for the client to get the friendlyName of a service without
a method call.