notices: only hide 'scrappedHours' notice after the day would be over without scrapped hours
For example, if all the hours on a day are scrapped the 'scrappedHours' notice
would have not been shown, since the day is over immediately.

This patch adds an parameter to the `ScheduleFunctions.current-day-over`
function, which tells the function to count scrapped hours as if it were normal
hours. The parameter is only used in the 'scrappedHours' notice for now, thanks
to this patch.