notices: don't show homework of schoolwide appointments in 'task' notice
This is to not show irrelevant  things on the overview as homework.

I had the `'content.type': $ne: 'information'` line for this, but some teachers
actually set the thing as homework.

Although I think this is a good solution, I'm not entirely sure. It could be
that we actually want to show schoolwide stuff sometimes. They are schoolwide
for a reason, no?