messages: cache to be stored drafts in localStorage
this is useful when somebodies internet dies, so they don't lose the draft when
the browser has been closed before it had a chance to store the draft in the
database.

maybe we should remove the draft from localStorage after successful insertion,
since currently we needlessly update already updated drafts on the next page
load.