external-services-connector: improve removing of `fetchServiceUpdates`
Before this commit calling `fetchServiceUpdates` would remove all the old
serviceUpdates before inserting the new ones, this is fine if all the services
are correctly working but if a service is down it would remove the notices from
that service too, while the notice could still exist on the service.

This commit fixes it by only removing the notices of the service that has just
been fetched, when an error occured it skips the removal and insertion of the
notices.