notices: run messages and grade cron jobs synchronously
This is practically a revert of 52aec4786dd70a92ad74920d4270907d52923c08,
although sending the emails themselves are still done concurrently, but with
this commit we're only handling the grades/messages of one user at a time.

This should hopefully limit CPU usage which went through the roof with
52aec4786dd70a92ad74920d4270907d52923c08, it did really improved the time it
took to run the cron jobs though. (from ~13m to ~3m (!))